+++ database/migrations/2020_06_05_141122_create_fav_artist_table.php
@@ -15,8 +15,8 @@ class CreateFavArtistTable extends Migration
     {
         Schema::create('fav_artist', function (Blueprint $table) {
             $table->bigIncrements('id');
-            $table->integer('user_id');
-            $table->integer('artist_id');
+            $table->unsignedBigInteger('user_id');
+            $table->unsignedBigInteger('artist_id');
             $table->timestamps();
         });
     }
